// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/StudentCreate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StudentCreate
 */
export interface StudentCreate {
    /**
     * 
     * @type {string}
     * @memberof StudentCreate
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof StudentCreate
     */
    birthday?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof StudentCreate
     */
    school?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StudentCreate
     */
    currentStageId: string;
    /**
     * 
     * @type {string}
     * @memberof StudentCreate
     */
    currentGradeId: string;
    /**
     * 
     * @type {string}
     * @memberof StudentCreate
     */
    remark?: string | null;
}

/**
 * Check if a given object implements the StudentCreate interface.
 */
export function instanceOfStudentCreate(value: object): value is StudentCreate {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('currentStageId' in value) || value['currentStageId'] === undefined) return false;
    if (!('currentGradeId' in value) || value['currentGradeId'] === undefined) return false;
    return true;
}

export function StudentCreateFromJSON(json: any): StudentCreate {
    return StudentCreateFromJSONTyped(json, false);
}

export function StudentCreateFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudentCreate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'birthday': json['birthday'] == null ? undefined : (new Date(json['birthday'])),
        'school': json['school'] == null ? undefined : json['school'],
        'currentStageId': json['currentStageId'],
        'currentGradeId': json['currentGradeId'],
        'remark': json['remark'] == null ? undefined : json['remark'],
    };
}

export function StudentCreateToJSON(json: any): StudentCreate {
    return StudentCreateToJSONTyped(json, false);
}

export function StudentCreateToJSONTyped(value?: StudentCreate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'birthday': value['birthday'] == null ? undefined : ((value['birthday'] as any).toISOString().substring(0,10)),
        'school': value['school'],
        'currentStageId': value['currentStageId'],
        'currentGradeId': value['currentGradeId'],
        'remark': value['remark'],
    };
}