// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/StudentUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StudentUpdate
 */
export interface StudentUpdate {
    /**
     * 
     * @type {string}
     * @memberof StudentUpdate
     */
    name: string;
    /**
     * 
     * @type {Date}
     * @memberof StudentUpdate
     */
    birthday?: Date;
    /**
     * 
     * @type {string}
     * @memberof StudentUpdate
     */
    school?: string;
    /**
     * 
     * @type {string}
     * @memberof StudentUpdate
     */
    currentStageId: string;
    /**
     * 
     * @type {string}
     * @memberof StudentUpdate
     */
    currentGradeId: string;
    /**
     * 
     * @type {string}
     * @memberof StudentUpdate
     */
    remark?: string;
    /**
     * 
     * @type {number}
     * @memberof StudentUpdate
     */
    version: number;
}

/**
 * Check if a given object implements the StudentUpdate interface.
 */
export function instanceOfStudentUpdate(value: object): value is StudentUpdate {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('currentStageId' in value) || value['currentStageId'] === undefined) return false;
    if (!('currentGradeId' in value) || value['currentGradeId'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    return true;
}

export function StudentUpdateFromJSON(json: any): StudentUpdate {
    return StudentUpdateFromJSONTyped(json, false);
}

export function StudentUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudentUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'birthday': json['birthday'] == null ? undefined : (new Date(json['birthday'])),
        'school': json['school'] == null ? undefined : json['school'],
        'currentStageId': json['currentStageId'],
        'currentGradeId': json['currentGradeId'],
        'remark': json['remark'] == null ? undefined : json['remark'],
        'version': json['version'],
    };
}

export function StudentUpdateToJSON(json: any): StudentUpdate {
    return StudentUpdateToJSONTyped(json, false);
}

export function StudentUpdateToJSONTyped(value?: StudentUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'birthday': value['birthday'] == null ? undefined : ((value['birthday']).toISOString().substring(0,10)),
        'school': value['school'],
        'currentStageId': value['currentStageId'],
        'currentGradeId': value['currentGradeId'],
        'remark': value['remark'],
        'version': value['version'],
    };
}